requirements.txt
# PyTorch和相关库
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# 计算机视觉
opencv-python>=4.5.0
efficientnet_pytorch>=0.7.0
Pillow>=8.3.0
facenet-pytorch>=2.5.0
ultralytics>=8.0.0
mtcnn>=0.1.1

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 工具库
tqdm>=4.62.0
PyYAML>=5.4.0
tensorboard>=2.7.0
wandb>=0.12.0
albumentations>=1.1.0
timm>=0.6.0

# 系统和IO
psutil>=5.8.0
h5py>=3.1.0
joblib>=1.1.0
pathos>=0.2.8

# 开发工具
jupyter>=1.0.0
ipywidgets>=7.6.0
black>=21.0.0
flake8>=3.9.0
pytest>=6.2.0

# 其他
requests>=2.25.0
colorama>=0.4.4
termcolor>=1.1.0